Moved services.py to bin
